Lyrics: improve sanitizing, move to xl.lyrics
Before this commit:
* LyricsMania would show a script at the bottom
* LyricWiki would not break lines correctly at `<br/>`
