+++ plugins/lyricsmania/__init__.py
@@ -79,6 +79,9 @@ class LyricsMania(LyricSearchMethod):
             lyrics_body = lyrics_html.find_class('lyrics-body')[0]
             lyrics_body.remove(lyrics_body.get_element_by_id('video-musictory'))
             lyrics = re.sub('^\s+Lyrics to .+', '', lyrics_body.text_content())
+            lyrics = lyrics.replace('\t', '')
+            lyrics = self.remove_script(lyrics)
+            lyrics = self.remove_html_tags(lyrics)
         except Exception:
             raise LyricsNotFoundException
